Moved FormIntegrationTestCase and FormPerformanceTestCase to the Test namespace
