+++ src/Symfony/Component/Form/Test/FormIntegrationTestCase.php
@@ -1,0 +1,41 @@
+<?php
+
+/*
+ * This file is part of the Symfony package.
+ *
+ * (c) Fabien Potencier <fabien@symfony.com>
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
+namespace Symfony\Component\Form\Test;
+
+use Symfony\Component\Form\Forms;
+
+/**
+ * @author Bernhard Schussek <bschussek@gmail.com>
+ */
+abstract class FormIntegrationTestCase extends \PHPUnit_Framework_TestCase
+{
+    /**
+     * @var \Symfony\Component\Form\FormFactoryInterface
+     */
+    protected $factory;
+
+    protected function setUp()
+    {
+        if (!class_exists('Symfony\Component\EventDispatcher\EventDispatcher')) {
+            $this->markTestSkipped('The "EventDispatcher" component is not available');
+        }
+
+        $this->factory = Forms::createFormFactoryBuilder()
+            ->addExtensions($this->getExtensions())
+            ->getFormFactory();
+    }
+
+    protected function getExtensions()
+    {
+        return array();
+    }
+}
