+++ src/Symfony/Component/Form/Tests/CompoundFormPerformanceTest.php
@@ -14,7 +14,7 @@ namespace Symfony\Component\Form\Tests;
 /**
  * @author Bernhard Schussek <bschussek@gmail.com>
  */
-class CompoundFormPerformanceTest extends FormPerformanceTestCase
+class CompoundFormPerformanceTest extends \Symfony\Component\Form\Tests\FormPerformanceTestCase
 {
     /**
      * Create a compound form multiple times, as happens in a collection form
